Refactored admin.py into seperate folder with files Refactored the category admin so that it is more usable
