+++ api/app/signals/apps/signals/admin/area.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2022 Gemeente Amsterdam
+from django.contrib.gis.admin import OSMGeoAdmin
+
+
+class AreaAdmin(OSMGeoAdmin):
+    search_fields = ['name', 'code', '_type__name', '_type__code']
+    list_display = ['name', 'code', '_type']
+    list_filter = ['_type__code']
